split base-carousel
